refactor(bundle-source): Separate evasive transform
